Effect System Started
I made most of the base for the effect system. Currently in place is the system with the ability to add and remove effects. The effects internally have their own stuff done. LArgely being added and removal as well as some starts for leave conditions and whatnot. This is largely by delegates.

I am going to need to make an event manager. That will be a multi-week project in itself but critical to the entire project.