Further Advanced The Effect System
EffectableComponent:  Created as a base ActorComponent which can be affected by effects. it has some simple code that would be way more annoying to repeatedly copy paste for handling the effect system on components. Currently used in the Health and Stamina components. I am going to have to create a movement component that we use instead of what is currently used so I can implement deeper functionality.

BaseEffectStruct: This is the struct used as a parent to all other effect structs (Stamina, health, etc). It has one pure virtual function for updating the values of the struct. This is used by EffectableComponent to update the values being used every so often. Basically any EffectableComponent will call one of these child structs to get information about its effects and this struct contains the data of all effects combined.